moved implementation relating to moving geometry objects around into its own directory
