+++ include/nektar_interface/geometry_transport/geometry_transport.hpp
@@ -3,6 +3,8 @@
 
 #include "geometry_transport_2d.hpp"
 #include "geometry_transport_3d.hpp"
-//#include "shape_mapping.hpp"
+#include "remote_geom_2d.hpp"
+#include "remote_geom_3d.hpp"
+#include "shape_mapping.hpp"
 
 #endif
